+++ src/Kopeeer.Worker/TargetConflictResolution.cs
@@ -1,0 +1,8 @@
+namespace Kopeeer.Worker;
+
+public enum TargetConflictResolution
+{
+    Rename = 0,
+    Skip = 1,
+    Cancel = 2
+}